rm repository/cars.py, path to csv_files, clean imports
