+++ .gitignore
@@ -11,3 +11,4 @@ car_parking/src/repository_or
 car_parking/src/routes_or
 idea
 main_or.py
+**/.idea
