--- car_parking/csv_files/first
@@ -1,8 +1,0 @@
-Name,Total Payment Amount,Total Parking Time,enter_time,departure_time,license_plate,amount_paid,duration,status
-,,,2024-04-11 12:52:50,2024-04-11 13:20:18,STRING,9.155555555555555,0.4577777777777778,True
-,,,2024-04-11 13:20:24,2024-04-11 13:20:32,STRING,0.03888888888888889,0.0019444444444444444,True
-,,,2024-04-11 13:20:35,2024-04-11 13:21:53,STRING,0.42777777777777776,0.021388888888888888,True
-,,,2024-04-11 13:21:55,2024-04-11 13:22:07,STRING,0.06666666666666667,0.0033333333333333335,True
-username,,,,,,,,
-,9.688888888888888,,,,,,,
-,,0.48444444444444446,,,,,,
