+++ car_parking/csv_files/test2.csv
@@ -1,0 +1,8 @@
+Name,Total Payment Amount,Total Parking Time,enter_time,departure_time,license_plate,amount_paid,duration,status
+,,,2024-04-12 14:12:43,2024-04-12 14:12:55,B02222CT,0.09166666666666667,0.0030555555555555557,True
+,,,2024-04-12 14:34:12,2024-04-12 14:34:48,B02222CT,0.2916666666666667,0.009722222222222222,True
+,,,2024-04-12 14:35:22,2024-04-12 14:36:07,B02222CT,0.25,0.0125,True
+,,,2024-04-12 16:28:24,2024-04-12 16:28:49,B02222CT,0.13333333333333333,0.006666666666666667,True
+kostia,,,,,,,,
+,0.7666666666666667,,,,,,,
+,,0.03194444444444444,,,,,,
